+++ tests/t00091/t00091_a.cc
@@ -1,0 +1,17 @@
+#include <string>
+
+namespace clanguml::t00091 {
+struct B;
+
+template <class T> struct C;
+
+template <> struct C<std::string>;
+
+enum class D;
+
+struct R {
+    B *b;
+    C<int> *c;
+    D d;
+};
+} // namespace clanguml::t00091